Allow passing the password through the environment.
Adds the option '--passwd-in-env' for passing the password through the
PASSWD environment variable.

This is useful in cases where the user would like to provide credentials
in the environment while also allowing interactive authorisation for
entering an OTP, for example if the server closes stdin after receiving
the OTP.

Changes:

- Add option '--passwd-in-env'

Note: this option isn't implemented for Windows.

Signed-off-by: Adam Tuft <adam.s.tuft@gmail.com>